feat: HospitalService outsorced in own service class. Fix bugs
